Factory and Abstract Factory

+++ src/main/java/kz/javazhan/sigma_finance/domain/entities/Account.java
@@ -15,7 +15,7 @@ public abstract class Account{
     UUID id;
     String accountNumber;
     AccountType accountType;
-    double balance;
+    long balance;
     CurrencyTypeEnum currencyType;
     AccountStatusEnum status;
 }
